+++ Definitions.h
@@ -5,7 +5,8 @@
 //#define DLS                  //set date on march 26 1:55, to test daylightsaving
 #define T5_E_PAPER           //for compiling without display funtions....
 #define GPIO12_ACTIF        //if GPIO12 is used as wake up, standard GPIO12 function is not activated !!
-
+//#define USE_AUTO_OTA_UPDATE
+#define VERSION 0
 #define TIME_DELAY_FIRST_FIX 10 //10 navpvt messages alvorens start loggen
 #define TIME_DELAY_NEW_RUN 10 //int time_delay_new_run
 #define EPOCH_2022 1640995200 //start of the year 2022 1640995200
@@ -16,6 +17,10 @@
 #define M10_38400BD 4
 #define M9_9600BD 5
 #define M9_38400BD 6
+#define M8_115200BD 7
+#define M9_115200BD 8
+#define M10_115200BD 9
+
 #define AUTO_DETECT 0xFF
 /*Pin definitions**************************************************************************/
 #define SPI_MOSI 23
@@ -25,8 +30,10 @@
 #define SDCARD_CLK 14
 #define SDCARD_MOSI 15
 #define SDCARD_MISO 2
-#define WAKE_UP_GPIO 39          //default 39
-#define GPIO_NUM_xx GPIO_NUM_39   //default GPIO_NUM_39 type is no int
+#define GO_TO_SLEEP_GPIO 39          //default 39
+#define WAKE_UP_GPIO_NUM GPIO_NUM_39   //default GPIO_NUM_39 type is no int
+#define GO_TO_SLEEP_PULLDOWN 19  //same function as 39, but here default is pulldown
+#define HOLD_PIN 21   //pin for high side switch, is high until shutdown ready
 #define RXD2 32 //yellow is Tx Ublox, Beitian white is Tx
 #define TXD2 33 //green is Rx Ublox, Beitian green is Rx
 #define PIN_BAT 35 //adc for bat voltage
@@ -41,7 +48,7 @@
 #define VOLTAGE_100 4.15     //reading 100% bat, after end charging voltage drops to 4.15 V
 #define VOLTAGE_0 3.4       //reading 0% bat
 #define VOLTAGE_LOW 25      //reading 25% bat
-#define MINIMUM_VOLTAGE 3.2       // if lower then minimum_voltage, back to sleep.....was 3.1
+#define MINIMUM_VOLTAGE 3.2     // if lower then minimum_voltage, back to sleep.....was 3.2
 #define MINIMUM_VOLTAGE_CHANGE  0.1  //refresh screen if change > 0.01
 #define STARTVALUE_HIGHEST_READ 2300  //12 bit ADC, 4.2 V over voltagedivider
 #define MAXVALUE_HIGHEST_READ 2700